+++ website/docs/language/state/examples/encryption/sample.tf
@@ -1,9 +1,13 @@
+variable "passphrase" {
+  # Change passphrase to be at least 16 characters long:
+  default = "changeme!"
+}
+
 terraform {
   encryption {
     ## Step 1: Add the desired key provider:
     key_provider "pbkdf2" "mykey" {
-      # Change this to be at least 16 characters long:
-      passphrase = "changeme!"
+      passphrase = var.passphrase
     }
     ## Step 2: Set up your encryption method:
     method "aes_gcm" "new_method" {